Move position exportingPDF
Move exportingPDF to Util and add case exception
